+++ docs/whatsnew.rst
@@ -5,6 +5,8 @@ What's new
 v0.9
 * Implementation of RBAC: per-resource access control, audit trail
 * List and outcomes are now editable
+* User managemetn UI
+* Role and permissions managemnt UI
 
 v0.7